requirements.txt
apache-airflow[crypto,celery,postgres,hive,hdfs,jdbc]==1.8.1
celery[redis]==3.1.17
pytz
pyOpenSSL
ndg-httpsclient
pyasn1
SQLAlchemy==1.1.9
pyquery==1.2.17
cryptography